+++ lib/apa102_rbpi/strip.rb
@@ -1,12 +1,10 @@
 module Apa102Rbpi
   class Strip
-
-    require 'set'
-
-    attr_reader :mirrors, :head, :tail, :num_leds
+    attr_reader :head, :tail, :num_leds, :base, :mirrors
     attr_accessor :led_frame_rgb_offsets, :brightness
 
-    def initialize(len = base.num_leds, opts = {})
+    def initialize(len = Apa102Rbpi.base.num_leds, opts = {})
+      @base = Apa102Rbpi.base
       if len.is_a?(Array)
         @head = len[0]
         @tail = len[1]
@@ -22,8 +20,8 @@ module Apa102Rbpi
 
       @reverse = false
 
-      @led_frame_rgb_offsets = opts[:led_frame_rgb_offsets] || base.led_frame_rgb_offsets
-      @brightness = opts[:brightness] || base.brightness
+      @led_frame_rgb_offsets = opts[:led_frame_rgb_offsets] || @base.led_frame_rgb_offsets
+      @brightness = opts[:brightness] || @base.brightness
       @mirrors = Set.new
     end
 
@@ -35,50 +33,60 @@ module Apa102Rbpi
       @reverse
     end
 
-    def mirror(strip)
-      if strip.num_leds == @num_leds
-        @mirrors.add(strip)
-        strip.mirrors.add(self)
-      else
-        raise 'Strips must be of same length to be mirrored!'
+    def mirror(other_strip)
+      new_mirrors = @mirrors + other_strip.mirrors + Set.new([self, other_strip])
+      new_mirrors.each do |m|
+        m.mirrors += (new_mirrors - [m])
       end
     end
 
-    def show!
-      base.show!
+    def clear_mirrors
+      @mirrors.each do |m|
+        m.mirrors -= [self]
+      end
+
+      @mirrors.clear
     end
 
-    def set_pixel(pos, color, brightness = @brightness)
-      q = [self]
-      seen = Set.new([self])
-      frames = base.led_frames
-      while(substrip = q.pop)
-        substrip.mirrors.each do |mirror|
-          unless seen.include?(mirror)
-            q.push(mirror)
-            seen.add(mirror)
-          end
-        end
+    def show!
+      @base.show!
+    end
 
-        idx = if substrip.reversed?
-                4 * ((substrip.tail - pos) % base.num_leds)
-              else
-                4 * ((pos + substrip.head) % base.num_leds)
-              end
+    def set_pixel(pos, color, brightness = nil)
+      is_hex = if color.is_a?(Integer)
+                  true
+               elsif color.is_a?(Array)
+                 false
+               else
+                 raise 'Invalid color'
+               end
 
-        if color.is_a?(Integer)
-          frames[idx] = substrip.led_frame_hdr(brightness)
-          frames[idx + substrip.led_frame_rgb_offsets[:red]] = (color & 0xFF0000) >> 16
-          frames[idx + substrip.led_frame_rgb_offsets[:green]] = (color & 0x00FF00) >> 8
-          frames[idx + substrip.led_frame_rgb_offsets[:blue]] = (color & 0x0000FF)
-        elsif color.is_a?(Array)
-          frames[idx] = substrip.led_frame_hdr(brightness)
-          frames[idx + substrip.led_frame_rgb_offsets[:red]] = color[0]
-          frames[idx + substrip.led_frame_rgb_offsets[:green]] = color[1]
-          frames[idx + substrip.led_frame_rgb_offsets[:blue]] = color[2]
-        else
-          raise 'Invalid color'
+      set_pixel_helper(pos, color, is_hex, brightness || @brightness)
+      unless @mirrors.empty?
+        @mirrors.each do |strip|
+          strip.set_pixel_helper(pos, color, is_hex, brightness || strip.brightness)
         end
+      end
+    end
+
+    def set_pixel_helper(pos, color, is_hex, brightness)
+      led_frame_hdr = (brightness & 0b00011111) | 0b11100000
+      idx = if @reverse
+              4 * ((@tail - pos) % @base.num_leds)
+            else
+              4 * ((pos + @head) % @base.num_leds)
+            end
+
+      if is_hex
+        @base.led_frames[idx] = led_frame_hdr
+        @base.led_frames[idx + @led_frame_rgb_offsets[:red]] = (color & 0xFF0000) >> 16
+        @base.led_frames[idx + @led_frame_rgb_offsets[:green]] = (color & 0x00FF00) >> 8
+        @base.led_frames[idx + @led_frame_rgb_offsets[:blue]] = (color & 0x0000FF)
+      else
+        @base.led_frames[idx] = led_frame_hdr
+        @base.led_frames[idx + @led_frame_rgb_offsets[:red]] = color[0]
+        @base.led_frames[idx + @led_frame_rgb_offsets[:green]] = color[1]
+        @base.led_frames[idx + @led_frame_rgb_offsets[:blue]] = color[2]
       end
     end
 
@@ -109,14 +117,10 @@ module Apa102Rbpi
       show!
     end
 
-    def led_frame_hdr(brightness = @brightness)
-      (brightness & 0b00011111) | 0b11100000
-    end
+    protected
 
-    private
-
-    def base
-      Apa102Rbpi.base
+    def mirrors=(m)
+      @mirrors = m
     end
   end
 end